Updated functionality

+++ app/controllers/uploads_controller.rb
@@ -7,20 +7,30 @@ class UploadsController < ApplicationController
 
   def new
 		@file = Upload.new
-		@course_id = params[:course_id]
-		redirect_to icons_path unless not @course_id.nil? 
+		if Course.find(params[:id]).nil?
+			flash[:error] = 'Error with finding course.'
+			redirect_to root_path
+		else
+			@course_id = params[:course_id]
+			redirect_to icons_path unless not @course_id.nil? 
+		end
   end
 	
   def create
-		@course_id = params[:course_id]
-		@file = Upload.new(secure_params)
-		@file.update_attributes(:user_id => current_user.id)
-
-		if @file.save
-			redirect_to icon_path(@file.course.id), :notice => "The file '#{@file.title}' has been uploaded for '#{@file.course.title}'."
+		if Course.find(params[:id]).nil?
+			flash[:error] = 'Error with finding course.'
+			redirect_to root_path
 		else
-			flash[:error] = @file.errors.full_messages.to_sentence.humanize
-			redirect_to :back
+			@course_id = params[:course_id]
+			@file = Upload.new(secure_params)
+			@file.update_attributes(:user_id => current_user.id)
+
+			if @file.save
+				redirect_to icon_path(@file.course.id), :notice => "The file '#{@file.title}' has been uploaded for '#{@file.course.title}'."
+			else
+				flash[:error] = @file.errors.full_messages.to_sentence.humanize
+				redirect_to :back
+			end
 		end
   end
 
@@ -33,9 +43,14 @@ class UploadsController < ApplicationController
 	end
 
   def destroy
-		@file = Upload.find(params[:id])
-		@file.destroy
-		redirect_to session.delete(:return_to), :notice => "The file '#{@file.title}' has been deleted."
+		if Upload.find(params[:id]).nil?
+			flash[:error] = 'Error with finding upload.'
+			redirect_to root_path
+		else
+			@file = Upload.find(params[:id])
+			@file.destroy
+			redirect_to session.delete(:return_to), :notice => "The file '#{@file.title}' has been deleted."
+		end
   end
 
 	private
